Commit 3
I have completed the automated claim status update. it used FluentRule which has the full criteria of the claim process and what needs to be there, and the threshold for the amounts according to SA Labor law. It uses these rules to determine whether the claim is legit or absurd and incorrect. if it aligns with the criteria, it will default status to pending else it will default to rejected.